pyproject.toml
[project]
name = "ClothesMatching"
version = "0.0.1"
dependencies = [
    "flet",
    "pillow",
    "numpy",
    "opencv-python"
]
authors = [
    { name = "Ekaterina Mashchenko" },
]
description = ""
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Operating System :: Android",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

[project.urls]
Homepage = "https://github.com/wielkate/ClothesMatching"

[tool.flet]
product = "Clothes matcher"

[tool.flet.app]
path = "src/package_wielkate/main"

[tool.flet.splash]
color = "#000000" # --splash-color
dark_color = "#000000" # --splash-dark-color

[tool.flet.app.boot_screen]
show = true
message = "Preparing the app for its first launch…"

[tool.flet.app.startup_screen]
show = true
message = "Starting up the app…"